vl*ff insn

+++ riscv/insns/vlb_v.h
@@ -9,7 +9,7 @@ check_vstart{
 	case 16:
 		vector_for_each(idx){
 			check_v0hmask(idx);
-			WRITE_VRD_H(MMU.load_int8(RS1+idx*SEW/8), idx);
+			WRITE_VRD_H((int16_t)MMU.load_int8(RS1+idx*SEW/8), idx);
 		}
 		break;
 	default:
