+++ riscv/insns/vlbu_v.h
@@ -10,7 +10,7 @@ check_vstart{
 	case 16:
 		vector_for_each(idx){
 			check_v0hmask(idx);
-			WRITE_VRD_H(MMU.load_uint8(RS1+idx*SEW/8), idx);
+			WRITE_VRD_H((uint16_t)MMU.load_uint8(RS1+idx*SEW/8), idx);
 		}
 		break;
 	default:
