+++ riscv/insns/vlse_v.h
@@ -4,7 +4,7 @@ check_vstart{
 	case 16:
 		vector_for_each(idx){
 			check_v0hmask(idx);
-			WRITE_VRD_H(MMU.load_int16(RS1+idx*RS2), idx);
+			WRITE_VRD_H((int16_t)MMU.load_int16(RS1+idx*RS2), idx);
 		}
 		break;
 	default:
