+++ riscv/insns/vlxb_v.h
@@ -4,7 +4,7 @@ check_vstart{
 	case 16:
 		vector_for_each(idx){
 			check_v0bmask(idx);
-			WRITE_VRD_H(MMU.load_int8(RS1+VRS2.vh[idx]), idx);
+			WRITE_VRD_H((int16_t)MMU.load_int8(RS1+VRS2.vh[idx]), idx);
 		}
 		break;
 	default:
