+++ riscv/insns/vlxbu_v.h
@@ -4,7 +4,7 @@ check_vstart{
 	case 16:
 		vector_for_each(idx){
 			check_v0hmask(idx);
-			WRITE_VRD_H(MMU.load_uint8(RS1+VRS2.vh[idx]), idx);
+			WRITE_VRD_H((uint16_t)MMU.load_uint8(RS1+VRS2.vh[idx]), idx);
 		}
 		break;
 	default:
